+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.0.2 (2018-08-11)
 - Add support for requests to both v1 and v2 endpoints
 - Retry failed requests few times before throwing an exception. That's configurable via the config file's failed_calls key
+- Dispatch RequestFailed and ResponseReceived events
 
 ## 1.0.1 (2018-07-13)
 - Allow ended games to be filtered by date. Available only for v2 of the betsapi.com API